Widen cc::PaintShader record shader IDs to 64 bits
Record shader IDs were previously minted from a 32-bit counter, which
could roll over in long-lived renderer processes. When the counter
wrapped around, ID collisions could cause the service-side transfer
cache to reuse an incorrect cached picture or generate
kInvalidRecordShaderId, resulting in raster failures.

Widen RecordShaderId and the sequence counter to uint64_t to prevent
counter rollover. uint64_t serialization in PaintOpWriter and
PaintOpReader is split into two uint32_t fields to maintain 4-byte
alignment guarantees, and the transfer cache lookup helper is updated to
accept 64-bit entry IDs.

TAG=agy

Bug: 556272421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8408827
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1702187}

+++ cc/paint/paint_shader.h
@@ -45,7 +45,7 @@ class CC_PAINT_EXPORT PaintShader : public SkRefCnt {
     kShaderCount
   };
 
-  using RecordShaderId = uint32_t;
+  using RecordShaderId = uint64_t;
   static const RecordShaderId kInvalidRecordShaderId;
 
   // Scaling behavior dictates how a PaintRecord shader will behave. Use
